requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
requests==2.31.0
beautifulsoup4==4.12.2
playwright==1.40.0
crawl4ai==0.2.77
python-multipart==0.0.6
aiofiles==23.2.1
python-dotenv==1.0.0
lxml==4.9.3
aiohttp==3.9.1
googlesearch-python==1.3.0 